# swagger.yaml
swagger: "2.0"
info:
  description: "This is a swagger spec for APIs provided by superdentist.io"
  version: "1.0.0"
  title: "SuperDentist APIs"
  contact:
    email: "superdentist.admin@superdentist.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "superdentist.io"
basePath: "/api/sd"
...
schemes:
- "https"
- "wss"
paths:
...
  /v1/referral/create:
    post:
      tags:
       - "Referrals"
      summary: "Create a referral from GD to Specialist"
      description: "Create a Referral and send out emails"
      operationId: "CreateRefSpecialist"
      consumes:
         - multipart/form-data
         - application/json
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created clinic object"
        required: true
        schema:
          $ref: "#/definitions/ReferralDetails"
      responses:
        '200':
          $ref: '#/responses/CreateReferralResponse'
        '401':
          description: "Unauthorized: Bad request or authorization details"
        '500':
          description: "Internal server error occured"
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ReferralResponse:
    type: object
    properties:
      referralId:
        type: string
      document:
        type: array
        items:
          type: string
      fromPlaceId:
        type: string
      toPlaceId:
        type: string
      fromClinicName:
        type: string
      toClinicName:
        type: string
      fromClinicAddress:
        type: string
      toClinicAddress:
        type: string
      comments:
        type: array
        items:
          type: object
          properties:
            time:
              type: string
            comment:
              type: string
            createdBy:
              type: string
      status:
        type: object
        properties:
          gdStatus:
            type: string
          spStatus:
            type: string
      reasons:
        type: array
        items:
          type: string
      history:
        type: array
        items:
          type: string
      tooth:
        type: array
        items:
          type: string
      createOn:
        type: string
      modifiedOn:
        type: string
      patientEmail:
        type: string
      patientFirstName:
        type: string
      patientLastName:
        type: string
      patientPhone:
        type: string
      fromEmail:
        type: string
      toEmail:
        type: string
      isDirty:
        type: boolean
        default: False
  ReferralDetails:
    type: object
    properties:
      patient:
        type: object
        properties:
          firstName:
            type: string
          lastName:
            type: string
          email:
            type: string
          phone:
            type: string
      fromAddressId:
        type: string
      toAddressId:
        type: string
      toPlaceId:
        type: string
      status:
        type: object
        properties:
          gdStatus:
            type: string
          spStatus:
            type: string
      comments:
        type: array
        items:
          type: object
          properties:
            time:
              type: string
            comment:
              type: string
            createdBy:
              type: string
      reasons:
        type: array
        items:
          type: string
      history:
        type: array
        items:
          type: string
      tooth:
        type: array
        items:
          type: string
...
responses:
...
  CreateReferralResponse:
    description: On successful referral creation payload containing unique referral id is returned and emails are send to patient and clinic for whom referral is created
    schema:
      $ref: '#/definitions/ReferralResponse'